requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2

# LLM Services
google-generativeai>=0.3.0

# Vector Database
qdrant-client>=1.6.0

# Document Processing
pymupdf>=1.23.0
docling>=1.0.0

# Web Interface
streamlit>=1.28.0

# Search APIs
google-search-results>=2.4.2

# Configuration
python-dotenv>=0.19.0

# Utilities
numpy>=1.24.0